jewellery_scraper/requirements.txt
icrawler
Pillow
imagehash
requests
beautifulsoup4
